Rename modules for queries.
